Now storing and displaying gifts per gift list

+++ gyft_storage/force_migrate.sh
@@ -1,9 +1,11 @@
+set -euxo pipefail
+
 echo "Deleting migration files"
 find . -path "*/migrations/*.py" -not -name "__init__.py" -delete
 find . -path "*/migrations/*.pyc"  -delete
 echo "Backuping and removing database"
-rm -f db.sqlite3.backup
-mv db.sqlite3 db.sqlite3.backup
+[ -e db.sqlite3.backup ] && rm db.sqlite3.backup
+[ -e db.sqlite3 ] && mv db.sqlite3 db.sqlite3.backup
 echo "Creating and running migrations"
 python manage.py makemigrations
 python manage.py migrate
